requirements.txt
# LangGraph + LangChain
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# Core
pydantic>=2.0
pyyaml>=6.0
jinja2>=3.1

# Web dashboard
fastapi>=0.110
uvicorn>=0.27
sse-starlette>=2.0

# pyXCP for real ECU access
pyxcp>=0.22
pya2ldb>=0.5

# CAN DBC parsing (optional — regex fallback if missing)
cantools>=39.0

# RAG pipeline
qdrant-client>=1.7
sentence-transformers>=2.3

# Reporting
pytest>=8.0
pytest-asyncio>=0.24
pytest-html>=4.1
weasyprint>=61.0

# Utilities
rich>=13.0
httpx>=0.27
python-dotenv>=1.0
